Refactor package responseDTO
